Final round of refactors after code review
