+++ apps/studio/components/interfaces/BranchManagement/EmptyStates.tsx
@@ -1,59 +1,10 @@
-import { PermissionAction } from '@supabase/shared-types/out/constants'
 import type { Branch } from 'data/branches/branches-query'
-import { ExternalLink, GitBranch, Github } from 'lucide-react'
+import { Github } from 'lucide-react'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import { BranchSelector } from './BranchSelector'
 
-import ProductEmptyState from 'components/to-be-cleaned/ProductEmptyState'
-import { ButtonTooltip } from 'components/ui/ButtonTooltip'
-import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
-import { useAppStateSnapshot } from 'state/app-state'
+import { DocsButton } from 'components/ui/DocsButton'
 import { Button } from 'ui'
-
-export const BranchingEmptyState = () => {
-  const snap = useAppStateSnapshot()
-
-  const canCreateBranches = useCheckPermissions(PermissionAction.CREATE, 'preview_branches', {
-    resource: { is_default: false },
-  })
-
-  return (
-    <ProductEmptyState title="Database Branching">
-      <p className="text-sm text-light">
-        Create preview branches to experiment with changes to your database schema in a safe,
-        non-destructive environment.
-      </p>
-      <div className="flex items-center space-x-2 !mt-4">
-        <ButtonTooltip
-          disabled={!canCreateBranches}
-          icon={<GitBranch strokeWidth={1.5} />}
-          onClick={() => snap.setShowCreateBranchModal(true)}
-          tooltip={{
-            content: {
-              side: 'bottom',
-              text: !canCreateBranches
-                ? 'You need additional permissions to create branches'
-                : undefined,
-            },
-          }}
-        >
-          Create branch
-        </ButtonTooltip>
-
-        <Button type="default" icon={<ExternalLink strokeWidth={1.5} />} asChild>
-          <a
-            target="_blank"
-            rel="noreferrer"
-            href="https://supabase.com/docs/guides/platform/branching"
-          >
-            View the docs
-          </a>
-        </Button>
-      </div>
-    </ProductEmptyState>
-  )
-}
 
 export const PullRequestsEmptyState = ({
   url,
@@ -72,8 +23,6 @@ export const PullRequestsEmptyState = ({
   githubConnection?: any
   gitlessBranching: boolean
 }) => {
-  const router = useRouter()
-
   return (
     <div className="flex items-center flex-col justify-center w-full py-10">
       <p>No merge requests</p>
@@ -83,16 +32,13 @@ export const PullRequestsEmptyState = ({
       <div className="flex items-center space-x-2 mt-4">
         {githubConnection ? (
           <Button type="outline" asChild icon={<Github />}>
-            <Link passHref href={url} target="_blank" rel="noreferrer">
+            <a href={url} target="_blank" rel="noopener noreferrer">
               Create pull request
-            </Link>
+            </a>
           </Button>
         ) : (
-          <Button
-            type="outline"
-            onClick={() => router.push(`/project/${projectRef}/settings/integrations`)}
-          >
-            Connect to GitHub
+          <Button asChild type="outline">
+            <Link href={`/project/${projectRef}/settings/integrations`}>Connect to GitHub</Link>
           </Button>
         )}
         {gitlessBranching && (
@@ -121,15 +67,7 @@ export const PreviewBranchesEmptyState = ({
         non-destructive environment.
       </p>
       <div className="flex items-center space-x-2">
-        <Button type="default" iconRight={<ExternalLink size={14} />}>
-          <Link
-            target="_blank"
-            rel="noreferrer"
-            href="https://supabase.com/docs/guides/platform/branching"
-          >
-            Docs
-          </Link>
-        </Button>
+        <DocsButton href="https://supabase.com/docs/guides/platform/branching" />
         <Button type="primary" onClick={() => onSelectCreateBranch()}>
           Create your first branch
         </Button>
