+++ apps/studio/components/layouts/AppLayout/BranchDropdown.tsx
@@ -2,7 +2,6 @@ import {
   AlertCircle,
   Check,
   ChevronsUpDown,
-  GitMerge,
   ListTree,
   MessageCircle,
   Plus,
@@ -15,7 +14,6 @@ import { useState } from 'react'
 import { useParams } from 'common'
 import ShimmeringLoader from 'components/ui/ShimmeringLoader'
 import { Branch, useBranchesQuery } from 'data/branches/branches-query'
-import { useBranchUpdateMutation } from 'data/branches/branch-update-mutation'
 import { useSelectedProject } from 'hooks/misc/useSelectedProject'
 import { useAppStateSnapshot } from 'state/app-state'
 import {
@@ -35,8 +33,6 @@ import {
   cn,
 } from 'ui'
 import { sanitizeRoute } from './ProjectDropdown'
-import { useFlag } from 'hooks/ui/useFlag'
-import { toast } from 'sonner'
 
 const BranchLink = ({
   branch,
@@ -78,21 +74,12 @@ const BranchLink = ({
 export const BranchDropdown = () => {
   const router = useRouter()
   const { ref } = useParams()
-  const gitlessBranching = useFlag('gitlessBranching')
-  const projectDetails = useSelectedProject()
   const snap = useAppStateSnapshot()
+  const projectDetails = useSelectedProject()
+
+  const [open, setOpen] = useState(false)
 
   const projectRef = projectDetails?.parent_project_ref || ref
-
-  const { mutate: updateBranch, isLoading: isUpdatingBranch } = useBranchUpdateMutation({
-    onSuccess: (_, variables) => {
-      if (variables.requestReview) {
-        router.push(`/project/${ref}/merge`)
-      }
-      setOpen(false)
-      toast.success('Branch marked as ready for review')
-    },
-  })
 
   const {
     data: branches,
@@ -122,100 +109,9 @@ export const BranchDropdown = () => {
         ? [mainBranch].concat(restOfBranches ?? [])
         : restOfBranches ?? []
       : [defaultMainBranch]
+  const branchList = isBranchingEnabled ? sortedBranches ?? [] : [defaultMainBranch]
 
   const BRANCHING_GITHUB_DISCUSSION_LINK = 'https://github.com/orgs/supabase/discussions/18937'
-
-  const [open, setOpen] = useState(false)
-
-  const MenuContent = ({
-    branchList,
-    showSearch,
-  }: {
-    branchList: Branch[]
-    showSearch?: boolean
-  }) => (
-    <Command_Shadcn_>
-      {showSearch && <CommandInput_Shadcn_ placeholder="Find branch..." />}
-      <CommandList_Shadcn_>
-        {showSearch && <CommandEmpty_Shadcn_>No branches found</CommandEmpty_Shadcn_>}
-
-        <CommandGroup_Shadcn_>
-          <ScrollArea className="max-h-[210px] overflow-y-auto">
-            {branchList.map((branch) => (
-              <BranchLink
-                key={branch.id}
-                branch={branch}
-                isSelected={branch.id === selectedBranch?.id || branches?.length === 0}
-                setOpen={setOpen}
-              />
-            ))}
-          </ScrollArea>
-        </CommandGroup_Shadcn_>
-
-        <CommandSeparator_Shadcn_ />
-
-        <CommandGroup_Shadcn_>
-          <CommandItem_Shadcn_
-            className="cursor-pointer w-full"
-            onSelect={() => {
-              setOpen(false)
-              snap.setShowCreateBranchModal(true)
-            }}
-            onClick={() => {
-              setOpen(false)
-              snap.setShowCreateBranchModal(true)
-            }}
-          >
-            <div className="w-full flex items-center gap-2">
-              <Plus size={14} strokeWidth={1.5} />
-              <p>Create branch</p>
-            </div>
-          </CommandItem_Shadcn_>
-          <CommandItem_Shadcn_
-            className="cursor-pointer w-full"
-            onSelect={() => {
-              setOpen(false)
-              router.push(`/project/${ref}/branches`)
-            }}
-            onClick={() => setOpen(false)}
-          >
-            <Link href={`/project/${ref}/branches`} className="w-full flex items-center gap-2">
-              <ListTree size={14} strokeWidth={1.5} />
-              <p>Manage branches</p>
-            </Link>
-          </CommandItem_Shadcn_>
-        </CommandGroup_Shadcn_>
-
-        <CommandSeparator_Shadcn_ />
-
-        <CommandGroup_Shadcn_>
-          <CommandItem_Shadcn_
-            className="cursor-pointer w-full"
-            onSelect={() => {
-              setOpen(false)
-              window?.open(BRANCHING_GITHUB_DISCUSSION_LINK, '_blank')?.focus()
-            }}
-            onClick={() => setOpen(false)}
-          >
-            <Link
-              href={BRANCHING_GITHUB_DISCUSSION_LINK}
-              target="_blank"
-              onClick={() => {
-                setOpen(false)
-              }}
-              className="w-full flex gap-2"
-            >
-              <MessageCircle size={14} strokeWidth={1} className="text-muted mt-0.5" />
-              <div>
-                <p>Branching feedback</p>
-                <p className="text-lighter">Join Github Discussion</p>
-              </div>
-            </Link>
-          </CommandItem_Shadcn_>
-        </CommandGroup_Shadcn_>
-      </CommandList_Shadcn_>
-    </Command_Shadcn_>
-  )
 
   return (
     <>
@@ -254,10 +150,91 @@ export const BranchDropdown = () => {
               />
             </PopoverTrigger_Shadcn_>
             <PopoverContent_Shadcn_ className="p-0" side="bottom" align="start">
-              <MenuContent
-                branchList={isBranchingEnabled ? sortedBranches ?? [] : [defaultMainBranch]}
-                showSearch={isBranchingEnabled}
-              />
+              <Command_Shadcn_>
+                {isBranchingEnabled && <CommandInput_Shadcn_ placeholder="Find branch..." />}
+                <CommandList_Shadcn_>
+                  {isBranchingEnabled && (
+                    <CommandEmpty_Shadcn_>No branches found</CommandEmpty_Shadcn_>
+                  )}
+
+                  <CommandGroup_Shadcn_>
+                    <ScrollArea className="max-h-[210px] overflow-y-auto">
+                      {branchList.map((branch) => (
+                        <BranchLink
+                          key={branch.id}
+                          branch={branch}
+                          isSelected={branch.id === selectedBranch?.id || branches?.length === 0}
+                          setOpen={setOpen}
+                        />
+                      ))}
+                    </ScrollArea>
+                  </CommandGroup_Shadcn_>
+
+                  <CommandSeparator_Shadcn_ />
+
+                  <CommandGroup_Shadcn_>
+                    <CommandItem_Shadcn_
+                      className="cursor-pointer w-full"
+                      onSelect={() => {
+                        setOpen(false)
+                        snap.setShowCreateBranchModal(true)
+                      }}
+                      onClick={() => {
+                        setOpen(false)
+                        snap.setShowCreateBranchModal(true)
+                      }}
+                    >
+                      <div className="w-full flex items-center gap-2">
+                        <Plus size={14} strokeWidth={1.5} />
+                        <p>Create branch</p>
+                      </div>
+                    </CommandItem_Shadcn_>
+                    <CommandItem_Shadcn_
+                      className="cursor-pointer w-full"
+                      onSelect={() => {
+                        setOpen(false)
+                        router.push(`/project/${ref}/branches`)
+                      }}
+                      onClick={() => setOpen(false)}
+                    >
+                      <Link
+                        href={`/project/${ref}/branches`}
+                        className="w-full flex items-center gap-2"
+                      >
+                        <ListTree size={14} strokeWidth={1.5} />
+                        <p>Manage branches</p>
+                      </Link>
+                    </CommandItem_Shadcn_>
+                  </CommandGroup_Shadcn_>
+
+                  <CommandSeparator_Shadcn_ />
+
+                  <CommandGroup_Shadcn_>
+                    <CommandItem_Shadcn_
+                      className="cursor-pointer w-full"
+                      onSelect={() => {
+                        setOpen(false)
+                        window?.open(BRANCHING_GITHUB_DISCUSSION_LINK, '_blank')?.focus()
+                      }}
+                      onClick={() => setOpen(false)}
+                    >
+                      <a
+                        target="_blank"
+                        rel="noreferrer noopener"
+                        href={BRANCHING_GITHUB_DISCUSSION_LINK}
+                        onClick={() => setOpen(false)}
+                        className="w-full flex gap-2"
+                      >
+                        <MessageCircle size={14} strokeWidth={1} className="text-muted mt-0.5" />
+                        <div>
+                          <p>Branching feedback</p>
+                          <p className="text-lighter">Join Github Discussion</p>
+                        </div>
+                      </a>
+                    </CommandItem_Shadcn_>
+                  </CommandGroup_Shadcn_>
+                </CommandList_Shadcn_>
+              </Command_Shadcn_>
             </PopoverContent_Shadcn_>
           </Popover_Shadcn_>
         </>
