--- apps/studio/components/layouts/BranchLayout/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BranchLayout'
